Add List of TopicFilters when getting Retained messages
This is so that storage can fetch only the messages that it needs
based on the current subscription, and not all messages.

If the broker is backed by a database, keeping all messages in memory
doesn't make sense. It would be better to call up individual sets of
messages on demand.